commit_1

+++ src/main/java/com/ryzhov_andrey/crud/repository/SkillRepository.java
@@ -2,5 +2,5 @@ package com.ryzhov_andrey.crud.repository;
 
 import com.ryzhov_andrey.crud.model.Skill;
 
-public interface SkillRepository extends GenericRepository <Skill,Integer>{
+public interface SkillRepository extends GenericRepository <Skill,Long>{
 }
